dataScience/ingestapython/requirements.txt
mysql-connector-python
pandas
boto3
faker
python-dotenv